requirements.txt
chromadb~=1.0.7
pypdf~=5.4.0
langchain~=0.3.25

python-dotenv~=1.1.0
pdfplumber~=0.11.6
langchain-community~=0.3.23
sentence-transformers
langchain-core~=0.3.58
langchain-text-splitters~=0.3.8